requirements.txt
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pathlib2>=2.3.0